Crate casper_dao_macros

Attribute Macros

casper_contract_interface

Generates contracts’ no_mangle functions.

Derive Macros

CLTyped

Derives CLType boilerplate code on top of any struct.

Event

Derive events on top of any struct.

FromBytes

Derives FromBytes boilerplate code on top of any struct.

Instance

Derives Instanced boilerplate code on top of any struct.

Rule

Derives a validation rule boilerplate code on top of any struct.

ToBytes

Derives ToBytes boilerplate code on top of any struct.